[#71] refactor:히스토리 페이지 알림 컴포넌트 분리
