+++ src/components/History/History/HistoryContent.js
@@ -5,7 +5,6 @@ import * as S from './History.Style';
 import {
   getHistoryList,
   deleteHistoryList,
-  readHistoryItem,
   moveToScrap,
   clearHistoryList,
   undoHistoryList,
@@ -18,14 +17,7 @@ import { useMediaQuery } from 'react-responsive';
 import theme from '../../../theme';
 import MobileMenuModal from './MobileMenuModal';
 import { MobileDeleteModal, MobileMoveScrapModal } from './MobilePopUpModal';
-export const formatingDate = (date) => {
-  const newDate = new Date(date);
-  const month = newDate.getMonth() + 1 < 10 ? '0' + (newDate.getMonth() + 1) : newDate.getMonth() + 1;
-  const day = newDate.getDate() < 10 ? '0' + newDate.getDate() : newDate.getDate();
-  const hour = newDate.getHours() < 10 ? '0' + newDate.getHours() : newDate.getHours();
-  const minute = newDate.getMinutes() < 10 ? '0' + newDate.getMinutes() : newDate.getMinutes();
-  return `${month + '/' + day} - ${(hour === 0 ? '00' : hour) + ':' + (minute===0?'00':minute)}`;
-};
+import HistoryAlert from './HistoryAlert';
 const HistoryContent = ({isToggleOpen}) => {
   const { historyList, deleteHistoryResponse, readHistoryItemResponse, moveToScrapResponse, undoHistoryListResponse } =
     useSelector((state) => state.history);
@@ -120,9 +112,6 @@ const HistoryContent = ({isToggleOpen}) => {
       setCheckedList(checkedList.filter((mailId) => mailId !== id));
     }
   };
-  const clickMail = (id) => {
-    dispatch(readHistoryItem(id));
-  };
   const closePopUp = () => {
     setOpen(false);
   };
@@ -212,7 +201,6 @@ const HistoryContent = ({isToggleOpen}) => {
   }, [isMobileDeleteOpen, isMobileScrapOpen]);
   return (
     <>
-    
       <S.PageWrapper onClick={isMobile ? closeMobileMenu : null}>
       {!isMobile && <PopUp isOpen={isPopOpen} closePopUp={closePopUp} />}
         <S.Content isOpen={isPopOpen}>
@@ -270,56 +258,29 @@ const HistoryContent = ({isToggleOpen}) => {
             </S.MenuWrapper>
           </S.MenuList>
           <S.KeyWordAlertList>
-            {showList?.map((mail, id) =>
+            {showList?.map((mail) =>
               showList[showList.length - 1].id === mail.id ? (
-                <S.KeyWordAlert isRead={mail.isRead} key={mail.id} ref={refAlert} isToggleOpen={isToggleOpen}>
-                  <HistoryCheckBox
-                    onClick={(e) => selectMail(e, mail.id)}
-                    checked={checkedList.includes(mail.id) ? true : false}
-                    readOnly
+                <div ref={refAlert} key={mail.id}>
+                  <HistoryAlert
+                    isToggleOpen={isToggleOpen}
+                    selectMail={selectMail}
+                    checkedList={checkedList}
+                    isMobile={isMobile}
+                    mail={mail}
                   />
-                  <S.AlertContent>
-                    <S.AlertDetail>
-                      <S.Sender>{SITE_LIST[SITE_LIST.findIndex((site) => site.id === mail.site)].title}</S.Sender>
-                      {isMobile && <S.ReceiveDate>{formatingDate(mail.createdAt)}</S.ReceiveDate>}
-                    </S.AlertDetail>
-                    <S.AlertTitle href={mail.url} isRead={mail.isRead} target='_blank' onClick={(e) => clickMail(mail.id)} isToggleOpen={isToggleOpen}>
-                      {mail.title}
-                    </S.AlertTitle>
-                  </S.AlertContent>
-                  {!isMobile && (
-                      <S.AlertInfo>
-                        <S.MailBrowse>{mail.isRead ? '읽음' : '읽지않음'}</S.MailBrowse>
-                        <S.ReceiveDate>{formatingDate(mail.createdAt)}</S.ReceiveDate>
-                      </S.AlertInfo>
-                    )}
-                </S.KeyWordAlert>
+                </div>
+
               ) : (
-                <>
-                  <S.KeyWordAlert isRead={mail.isRead} key={mail.id} isToggleOpen={isToggleOpen}>
-                    <HistoryCheckBox
-                      onClick={(e) => selectMail(e, mail.id)}
-                      checked={checkedList.includes(mail.id) ? true : false}
-                      readOnly
-                    />
-                    <S.AlertContent>
-                      <S.AlertDetail>
-                        <S.Sender>{SITE_LIST[SITE_LIST.findIndex((site) => site.id === mail.site)].title}</S.Sender>
-                        {isMobile && <S.ReceiveDate>{formatingDate(mail.createdAt)}</S.ReceiveDate>}
-                      </S.AlertDetail>
-                      <S.AlertTitle href={mail.url} isRead={mail.isRead} target='_blank' onClick={(e) => clickMail(mail.id)} isToggleOpen={isToggleOpen}>
-                        {mail.title}
-                      </S.AlertTitle>
-                    </S.AlertContent>
-                    {!isMobile && (
-                      <S.AlertInfo>
-                        <S.MailBrowse>{mail.isRead ? '읽음' : '읽지않음'}</S.MailBrowse>
-                        <S.ReceiveDate>{formatingDate(mail.createdAt)}</S.ReceiveDate>
-                      </S.AlertInfo>
-                    )}
-                  </S.KeyWordAlert>
+                <div key={mail.id}>
+                 <HistoryAlert
+                  isToggleOpen={isToggleOpen}
+                  selectMail={selectMail}
+                  checkedList={checkedList}
+                  isMobile={isMobile}
+                  mail={mail}
+                />
                   {isMobile && <S.AlertBorderBox />}
-                </>
+                </div>
               )
             )}
           </S.KeyWordAlertList>
